# swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  version: 1.0.0
  description: API for user login and logout functionality
servers:
  - url: http://localhost:3000/api # Replace with your actual base API URL
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Allows users to log in with their gmail and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gmail:
                  type: string
                  description: User's gmail address
                password:
                  type: string
                  description: User's password
                  format: password
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: User's ID
                  token:
                    type: string
                    description: JWT authentication token
                  gmail:
                    type: string
                    description: User's gmail address
        401:
          description: Invalid username or password, or user disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid username or password
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Allows a logged-in user to log out and invalidate their session.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful logout
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully logged out!
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
...
  /friends:
    post:
      tags:
        - Friends
      summary: Create a new friend
      description: Creates a new friend and then create a new conversation between them.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gmail:
                  type: string
                  description: Friend's gmail address
      responses:
        201:
          description: Friend added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Bad request
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /friends/{id}:
    get:
      tags:
        - Friends
      summary: Get friends of a specific user
      description: Retrieves an array of friends of an user with userId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        200:
          description: Friends retrieved successfully
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/FriendInUser"
        404:
          description: Friends not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Friends not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

    delete:
      tags:
        - Friends
      summary: Delete a friend
      description: Delete a friend.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Friend ID
      responses:
        200:
          description: Friend deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
        404:
          description: Friend not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Friend not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Friend:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        friendId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FriendInUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        friendId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: integer
            gmail:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            middleName:
              type: string
            dateOfBirth:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            avatarName:
              type: string
            disabled:
              type: boolean
...